server/requirements.txt
Django==6.0
django-ninja==1.5.1
EbookLib==0.20
python-dotenv==1.2.1
requests==2.32.3
django-cors-headers==4.9.0
beautifulsoup4==4.14.3
textblob==0.19.0
redis==7.1.0
